chore: use same project resolving as the project picker
